pyproject.toml
[project]
name = "airas-experiments"
version = "0.1.0"
description = "AIRAS ML experiment framework"
requires-python = ">=3.11"
dependencies = [
    "hydra-core>=1.3.2",
    "wandb>=0.16.0",
    "torch>=2.1.0",
    "transformers>=4.36.0",
    "datasets>=2.16.0",
    "accelerate>=0.25.0",
    "pandas>=2.1.0",
    "numpy>=1.24.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "pyyaml>=6.0",
    "regex>=2023.0.0",
]
